+++ src/main/java/com/tacs/backend/dto/PlayerStatisticsDTO.java
@@ -13,5 +13,5 @@ import lombok.Setter;
 @Schema(description = "Players statistics")
 public class PlayerStatisticsDTO {
   @Schema(description = "Number of players enrolled in the last two hours")
-  int playersEnrolled;
+  long playersEnrolled;
 }
